pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copilot-ansible-agent"
version = "0.1.0"
description = "Copilot Studio AI agent for Ansible automation"
requires-python = ">=3.11"
authors = [
    { name = "Your Team", email = "team@example.com" }
]
dependencies = [
    "pydantic>=2.5",
    "paramiko>=3.4",
    "PyYAML>=6.0",
    "rich>=13.7",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1.8",
    "mypy>=1.7"
]

[project.scripts]
copilot-ansible-agent = "copilot_ansible_agent.__main__:main"
